Separate book upload and download into their own classes
