+++ src/BloomExe/WebLibraryIntegration/AccessKeys.cs
@@ -48,7 +48,7 @@ namespace Bloom.WebLibraryIntegration
 				case BloomS3Client.SandboxBucketName:
 					// S3 'uploaderDev' user, who has permission to use the BloomLibraryBooks-Sandbox bucket.
 					//parse.com silbloomlibrarysandbox
-					if (BookTransfer.IsDryRun)
+					if (BookUpload.IsDryRun)
 						return new AccessKeys(null, null, lines[6], lines[7]);
 					return new AccessKeys(lines[2], lines[3],lines[6],lines[7]);
 				case BloomS3Client.UnitTestBucketName:
@@ -56,7 +56,7 @@ namespace Bloom.WebLibraryIntegration
 				case BloomS3Client.ProductionBucketName:
 					//S3 'uploader' user, who has permission to use the BloomLibraryBooks bucket
 					//parse.com silbloomlibrary
-					if (BookTransfer.IsDryRun)
+					if (BookUpload.IsDryRun)
 						return new AccessKeys(null, null, lines[4], lines[5]);
 					return new AccessKeys(lines[0], lines[1], lines[4], lines[5]);
 				case BloomS3Client.ProblemBookUploadsBucketName:
