+++ src/BloomExe/WebLibraryIntegration/BloomParseClient.cs
@@ -22,7 +22,7 @@ namespace Bloom.WebLibraryIntegration
 
 		public BloomParseClient()
 		{
-			var keys = AccessKeys.GetAccessKeys(BookTransfer.UploadBucketNameForCurrentEnvironment);
+			var keys = AccessKeys.GetAccessKeys(BookUpload.UploadBucketNameForCurrentEnvironment);
 
 			RestApiKey = keys.ParseApiKey;
 			ApplicationId = keys.ParseApplicationKey;
@@ -107,7 +107,7 @@ namespace Bloom.WebLibraryIntegration
 
 		public string GetRealUrl()
 		{
-			return UrlLookup.LookupUrl(UrlType.Parse, BookTransfer.UseSandbox);
+			return UrlLookup.LookupUrl(UrlType.Parse, BookUpload.UseSandbox);
 		}
 
 		private RestRequest MakeRequest(string path, Method requestType)
@@ -243,7 +243,7 @@ namespace Bloom.WebLibraryIntegration
 		{
 			if (!LoggedIn)
 				throw new ApplicationException();
-			if (BookTransfer.IsDryRun)
+			if (BookUpload.IsDryRun)
 				throw new ApplicationException("Should not call CreateBookRecord during dry run!");
 			var request = MakePostRequest("classes/books");
 			request.AddParameter("application/json", metadataJson, ParameterType.RequestBody);
@@ -266,7 +266,7 @@ namespace Bloom.WebLibraryIntegration
 		{
 			if (!LoggedIn)
 				throw new ApplicationException("BloomParseClient got SetBookRecord, but the user is not logged in.");
-			if (BookTransfer.IsDryRun)
+			if (BookUpload.IsDryRun)
 				throw new ApplicationException("Should not call SetBookRecord during dry run!");
 			var metadata = BookMetaData.FromString(metadataJson);
 			var book = GetSingleBookRecord(metadata.Id);
@@ -297,7 +297,7 @@ namespace Bloom.WebLibraryIntegration
 		{
 			if (!LoggedIn)
 				throw new ApplicationException("Must be logged in to delete book");
-			if (BookTransfer.IsDryRun)
+			if (BookUpload.IsDryRun)
 				throw new ApplicationException("Should not call DeleteBookRecord during dry run!");
 			var request = MakeDeleteRequest("classes/books/" + bookObjectId);
 			var response = Client.Execute(request);
@@ -309,7 +309,7 @@ namespace Bloom.WebLibraryIntegration
 		{
 			if (!LoggedIn)
 				throw new ApplicationException();
-			if (BookTransfer.IsDryRun)
+			if (BookUpload.IsDryRun)
 				throw new ApplicationException("Should not call DeleteLanguages during dry run!");
 			var getLangs = MakeGetRequest(ClassesLanguagePath);
 			var response1 = Client.Execute(getLangs);
@@ -329,7 +329,7 @@ namespace Bloom.WebLibraryIntegration
 		{
 			if (!LoggedIn)
 				throw new ApplicationException();
-			if (BookTransfer.IsDryRun)
+			if (BookUpload.IsDryRun)
 			{
 				Console.WriteLine("Simulating CreateLanguage during dry run for {0} ({1})", lang.Name, lang.EthnologueCode);
 				return JObject.Parse($"{{\"objectId\":\"xyzzy{lang.EthnologueCode}\"}}");
@@ -395,7 +395,7 @@ namespace Bloom.WebLibraryIntegration
 
 		internal void SendResetPassword(string account)
 		{
-			if (BookTransfer.IsDryRun)
+			if (BookUpload.IsDryRun)
 				throw new ApplicationException("Should not call SendResetPassword during dry run!");
 			var request = MakePostRequest("requestPasswordReset");
 			request.AddParameter("application/json; charset=utf-8", "{\"email\":\""+account+ "\"}", ParameterType.RequestBody);
